Add parameter types
This is a breaking change since consumers may be providing entities that can be converted to a string.

Implementations that implement v2, however, are also implementing v1 as of PHP 7.2, so such implementations can use a require statement of `^1|^2`. Usage libraries also can use a require statement of `^1|^2` so long as they are passing the specific variable type